pyproject.toml
[project]
name = "flaskintro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "mysql-connector-python>=9.6.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "scikit-learn>=1.8.0",
]
